requirements.txt
faker
sqlalchemy
polars
python-dotenv